worker_function/requirements.txt
google-cloud-firestore>=2.17.0


